+++ net-spider-rpl/src/NetSpider/RPL/JSONUtil.hs
@@ -1,0 +1,19 @@
+-- |
+-- Module: NetSpider.RPL.JSONUtil
+-- Description: 
+-- Maintainer: Toshio Ito <debug.ito@gmail.com>
+--
+-- __This is an internal module. End-users should not use it.__
+module NetSpider.RPL.JSONUtil
+  ( optSnake
+  ) where
+
+import qualified Data.Aeson as Aeson
+import Data.Char (isUpper, toLower)
+import qualified Text.Regex.Applicative as RE
+
+toSnake :: String -> String
+toSnake = RE.replace $ RE.msym (\c -> if isUpper c then Just ['_', toLower c] else Nothing)
+
+optSnake :: Aeson.Options
+optSnake = Aeson.defaultOptions { Aeson.fieldLabelModifier = toSnake }